switched to fragments and deleted activities
