+++ .idea/misc.xml
@@ -10,9 +10,12 @@
         <entry key="app/src/main/res/drawable/circle.xml" value="0.2722222222222222" />
         <entry key="app/src/main/res/layout/activity_full_image.xml" value="0.3458333333333333" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.36875" />
+        <entry key="app/src/main/res/layout/activity_photos.xml" value="0.3458333333333333" />
         <entry key="app/src/main/res/layout/activity_splash.xml" value="0.38385416666666666" />
         <entry key="app/src/main/res/layout/album_view.xml" value="0.36875" />
+        <entry key="app/src/main/res/layout/fragment_full_image.xml" value="0.3458333333333333" />
         <entry key="app/src/main/res/layout/fragment_main.xml" value="0.353125" />
+        <entry key="app/src/main/res/layout/fragment_scrolling.xml" value="0.3458333333333333" />
         <entry key="app/src/main/res/layout/photo_view.xml" value="0.36875" />
         <entry key="app/src/main/res/layout/scrolling_activity.xml" value="0.38385416666666666" />
         <entry key="app/src/main/res/layout/style.xml" value="0.3458333333333333" />
