+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Gallery">
         <activity
-            android:name=".ScrollingActivity"
+            android:name=".MainActivity"
             android:exported="true">
 
             <intent-filter>
@@ -24,7 +24,7 @@
 
         </activity>
 
-        <activity android:name=".FullImageActivity" />
+<!--        <activity android:name="" />-->
         <activity
             android:name=".PhotosActivity"
             android:exported="true">
