+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">Gallery</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="share">Share</string>
 </resources>
